Move worker config from admin to default
